+++ src/app/index.ts
@@ -2,8 +2,3 @@
  * app module
  */
 export { appModule } from './app.module'
-
-/**
- * app constants
- */
-export { default as appConstants } from './app.constants'