+++ webpack.config.js
@@ -1,20 +1,40 @@
+// core modules
+const path = require('path')
+
 // npm modules
-const ip = require('ip')
-
-// Webpack plugins
 const HtmlWebpackPlugin = require('html-webpack-plugin')
+const ip = require('ip')
+const merge = require('webpack-merge')
+const validate = require('webpack-validator')
 
-// Config parts
-const webpackParts = require('./libs/webpack.parts')
+// Configuration parts
+const parts = require('./libs/webpack.parts')
 
-/**
- * Webpack configuration for development environment
- */
-module.exports = {
-  context: __dirname,
-  entry: './src/main.ts',
+// Constants & Variables
+const PATHS = {
+  app: path.join(__dirname, 'src'),
+  dist: path.join(__dirname, 'dist'),
+  images: path.join(__dirname, 'src/assets/img'),
+  fonts: path.join(__dirname, 'src/assets/fonts'),
+  copy: [
+    {
+      from: './src/assets',
+      to: 'assets'
+    }, {
+      from: './src/favicon.ico'
+    }, {
+      from: './src/robots.txt'
+    }]
+}
+
+// Common settings for webpack
+const common = {
+  entry: {
+    app: PATHS.app
+  },
   output: {
-    filename: 'main.js'
+    path: PATHS.dist,
+    filename: '[name].js'
   },
   plugins: [
     new HtmlWebpackPlugin({
@@ -22,9 +42,7 @@ module.exports = {
       hash: true
     })
   ],
-  devtool: 'source-map',
   resolve: {
-    root: __dirname,
     extensions: ['', '.ts', '.js', '.json', '.css']
   },
   resolveLoader: {
@@ -45,50 +63,72 @@ module.exports = {
       {
         test: /\.ts$/,
         loader: 'ts',
-        exclude: /node_modules/
+        include: PATHS.app
       },
       {
         test: /\.html$/,
-        exclude: /\index.html/,
-        loader: 'html'
-      },
-      {
-        test: /\.css$/,
-        loader: 'style!css!postcss'
+        loader: 'html',
+        include: PATHS.app
       },
       {
         test: /\.json$/,
-        loader: 'json'
+        loader: 'json',
+        include: PATHS.app
       },
       {
         test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-        loader: 'url?limit=10000&mimetype=application/font-woff'
+        loader: 'url?limit=10000&mimetype=application/font-woff',
+        include: PATHS.fonts
       },
       {
         test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-        loader: 'file'
+        loader: 'file',
+        include: PATHS.fonts
       },
       {
         test: /\.jpg$/,
-        exclude: /node_modules/,
-        loader: 'file'
+        loader: 'file',
+        include: PATHS.images
       },
       {
         test: /\.png$/,
-        exclude: /node_modules/,
-        loader: 'url'
+        loader: 'url',
+        include: PATHS.images
       }
     ]
   },
-  tslint: webpackParts.tslint,
-  postcss: webpackParts.postcss,
-  devServer: {
-    host: process.env.ANGULARTS_WP_HOST || ip.address(),
-    port: process.env.ANGULARTS_WP_PORT || 3000,
-    contentBase: ['./src', '../static'],
-    historyApiFallback: true,
-    inline: true,
-    colors: true,
-    stats: 'normal'
-  }
+  tslint: parts.tslint,
+  postcss: parts.postcss
 }
+
+// Detect how npm is run and branch based on that
+let config
+switch (process.env.npm_lifecycle_event) {
+  case 'build':
+    config = merge(
+      common,
+      parts.clean(PATHS.dist),
+      parts.minify(),
+      parts.extractCSS(PATHS.app),
+      parts.copy(PATHS.copy)
+    )
+    break
+  default:
+    config = merge(
+      common,
+      {
+        devtool: 'eval-source-map'
+      },
+      parts.setupCSS(PATHS.app),
+      parts.devServer({
+        // Customize host/port here if needed
+        host: process.env.ANGULARTS_WP_HOST || ip.address(),
+        port: process.env.ANGULARTS_WP_PORT || 3000
+      })
+    )
+}
+
+/**
+ * Webpack configuration
+ */
+module.exports = validate(config)